requirements.txt
python==3.5.4
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.2
cryptography==2.1.4
Django==2.0
django-bootstrap3==9.1.0
django-popup-view-field==0.3.0
Fabric3==1.13.1.post1
idna==2.6
paramiko==2.4.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2017.3
six==1.11.0
